requirements.txt
ollama==0.1.7
streamlit==1.39.0
langchain==1.1.7
langchain-core==1.1.20
langchain-community==0.3.3
python-dotenv==1.0.1
requests==2.32.3
typing-extensions==4.12.2
